+++ CHANGES.md
@@ -43,6 +43,7 @@
 - FIX: Fixing the band mapping of `WorldView-2/3 Multi` (8 bands)
 - FIX: Retrieval (if possible) of Sentinel-1 [unique ID](https://sentinels.copernicus.eu/web/sentinel/user-guides/sentinel-1-sar/naming-conventions) (was missing from the product name, as it is not in the product preview)
 - FIX: Fixing PAZ/TDX MTD regex
+- FIX: Optical products: Only set cloud cover and radiometry attributes if spectral bands are asked
 
 ### Other